text_2_sql/resumedb/database_app/requirements.txt
absl-py==1.4.0
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
appdirs==1.4.4
asgiref==3.7.2
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
awscrt==0.16.26
backcall==0.2.0
bertopic==0.15.0
bitsandbytes==0.39.1
black==23.7.0
bleach==6.0.0
boto3==1.28.17
botocore==1.31.17
Brotli==1.0.9
brotlipy==0.7.0
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==2.0.4
click==8.1.6
cmake==3.27.0
comm==0.1.3
cryptography==39.0.1
Cython==0.29.36
datasets==2.13.1
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
Django==4.2.3
django-cors-headers==4.2.0
djangorestframework==3.14.0
executing==1.2.0
fairscale==0.4.13
filelock==3.9.0
fire==0.5.0
frozenlist==1.4.0
fsspec==2023.6.0
gmpy2==2.1.2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
greenlet==2.0.2
grpcio==1.56.2
hdbscan==0.8.33
httpie==3.2.2
huggingface-hub==0.16.4
idna==3.4
inflate64==0.3.1
ipykernel==6.24.0
ipython==8.14.0
ipywidgets==8.0.7
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.3.1
jupyter_client==8.3.0
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kaggle==1.5.16
lit==16.0.6
llvmlite==0.40.1
loralib==0.1.1
Markdown==3.4.3
markdown-it-py==3.0.0
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.4
multiprocess==0.70.14
multivolumefile==0.2.3
mypy-extensions==1.0.0
nest-asyncio==1.5.6
networkx==2.8.4
nltk==3.8.1
numba==0.57.1
numpy==1.24.4
oauthlib==3.2.2
packaging==23.1
pandas==2.0.3
parso==0.8.3
pathspec==0.11.1
peft==0.5.0.dev0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pip==23.1.2
platformdirs==3.9.1
plotly==5.15.0
prompt-toolkit==3.0.39
protobuf==3.20.0
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py7zr==0.20.5
pyarrow==12.0.1
pyasn1==0.5.0
pyasn1-modules==0.3.0
pybcj==1.0.1
pycparser==2.21
pycryptodomex==3.18.0
Pygments==2.15.1
pynndescent==0.5.10
pyOpenSSL==23.0.0
pyppmd==1.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytz==2023.3
PyYAML==6.0.1
pyzmq==25.1.0
pyzstd==0.15.9
rapidfuzz==3.1.2
regex==2023.6.3
requests==2.29.0
requests-oauthlib==1.3.1
requests-toolbelt==1.0.0
rich==13.5.0
rsa==4.9
s3transfer==0.6.1
safetensors==0.3.1
scikit-learn==1.3.0
scipy==1.11.1
sentence-transformers==2.2.2
sentencepiece==0.1.99
setuptools==67.8.0
six==1.16.0
SQLAlchemy==2.0.19
sqlglot==17.9.1
sqlparse==0.4.4
stack-data==0.6.2
sympy==1.11.1
tenacity==8.2.2
tensorboard==2.13.0
tensorboard-data-server==0.7.1
termcolor==2.3.0
text-unidecode==1.3
texttable==1.6.7
threadpoolctl==3.2.0
tokenize-rt==5.1.0
tokenizers==0.13.3
torch==2.0.1+cu118
torch-tb-profiler==0.4.1
torchaudio==2.0.2
torchvision==0.15.2
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.32.0.dev0
triton==2.0.0
trl==0.4.7
typing_extensions==4.6.3
tzdata==2023.3
umap-learn==0.5.3
urllib3==1.26.16
wcwidth==0.2.6
webencodings==0.5.1
Werkzeug==2.3.6
wheel==0.38.4
widgetsnbextension==4.0.8
xxhash==3.2.0
yarl==1.9.2
